+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added chat frame customization
 -   Added experience bar customization
 -   New miscellanous that takes screenshots when achievements are earned
+-   New miscellanous to track skyriding race timer
 
 ### Changed